feat(terraform-modules/azure-nat-gateway) introduce support for additional outbound IPs
This change adds the ability to increase the amount of outbound IPs.

Note that the addition IPs are specified as a distinct resource set than the current `azurerm_public_ip.outbound`:

- The "multiple outbound IPs" is an edge case for trusted.ci (at least for now)
- The idea is to avoid having to move resources inside terraform state to go quickly.

=> It is only an implementation detail: the input variables and module outputs are mixing
the "principal" outbound IP and the eventual addition all together so consumers don't have to care.

=> If the reality catch us and makes the "multiple outbound IPs" the nominal case, then we'll consider
implementing a single set of azurerm_public_ip and migrate existing resources once for all.

Signed-off-by: Damien Duportal <damien.duportal@gmail.com>